// src/types/CampagnePromoStats.ts
export interface CampagnePromoStats {
    idCampagne: string;
    nombreCommandes: number;
    totalArticlesVendus: number;
    chiffreAffairesReel: number;
    chiffreAffairesTheorique: number;
    manqueAGagner: number;
    articlesVendusAvantPromo: number;
    pourcentageAugmentationVolume: number;
    chiffreAffairesAvantPromo: number;
    pourcentageAugmentationRevenu: number;
}